requirements.txt
absl-py==0.1.9
alembic==1.0.0
asn1crypto==0.23.0
astor==0.7.1
attrs==17.3.0
Automat==0.6.0
autopep8==1.4.2
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
blinker==1.4
bs4==0.0.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cos-python-sdk-v5==1.5.8
cryptography==2.1.3
cssselect==1.0.1
cycler==0.10.0
decorator==4.1.2
dicttoxml==1.7.4
dominate==2.3.1
entrypoints==0.2.3
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
ForgeryPy==0.1
futures==3.1.1
gast==0.2.2
grpcio==1.18.0
h5py==2.8.0
html5lib==0.9999999
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipykernel==4.6.1
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.0.4
itsdangerous==0.24
jedi==0.11.0
jieba==0.39
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.0
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.6
lxml==4.1.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.1
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
parsel==1.2.0
parso==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.6.1
pycodestyle==2.4.0
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pymongo==3.6.0
PyMySQL==0.9.2
pyOpenSSL==17.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.9.0
selenium==3.8.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.11
tensorboard==1.12.2
tensorflow-gpu==1.12.0
tensorflow-hub==0.2.0
tensorflow-tensorboard==1.5.0
termcolor==1.1.0
terminator==0.0.0
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
urllib3==1.22
visitor==0.1.3
w3lib==1.18.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.0.7
WTForms==2.2.1
zhon==1.1.5
zope.interface==4.4.3
